chore: Move `LoggingConfig` into `cda-config`.
