+++ cda-config/Cargo.toml
@@ -36,3 +36,7 @@ workspace = true
 
 [lib]
 path = "src/lib.rs"
+
+[features]
+tokio-tracing = []
+dlt-tracing = []
