+++ cda-tracing/src/lib.rs
@@ -12,8 +12,8 @@
  */
 use std::ops::Deref;
 
+use cda_config::datatypes::{LogFileConfig, LoggingConfig, OtelConfig};
 use opentelemetry::trace::TracerProvider;
-use serde::{Deserialize, Serialize};
 use thiserror::Error;
 use tracing_appender::non_blocking::WorkerGuard;
 use tracing_subscriber::{
@@ -24,11 +24,8 @@ use tracing_subscriber::{
 };
 
 mod otel;
-pub use otel::{OtelConfig, OtelGuard};
+pub use otel::OtelGuard;
 pub mod subscriber;
-
-const DEFAULT_LOG_FILE_NAME: &str = "opensovd-cda.log";
-const DEFAULT_LOG_FILE_PATH: &str = "/var/log/opensovd-cda";
 
 #[derive(Error, Debug)]
 pub enum TracingSetupError {
@@ -229,7 +226,7 @@ pub fn new_otel_subscriber<
 /// Returns an error if the socket address cannot be parsed.
 #[cfg(feature = "tokio-tracing")]
 pub fn new_tokio_tracing<S: tracing_core::Subscriber + for<'a> LookupSpan<'a>>(
-    config: &TokioTracingConfig,
+    config: &cda_config::datatypes::TokioTracingConfig,
 ) -> Result<BoxedLayer<S>, TracingSetupError> {
     use std::net::SocketAddr;
 
@@ -253,7 +250,7 @@ pub fn new_tokio_tracing<S: tracing_core::Subscriber + for<'a> LookupSpan<'a>>(
 /// The app id is limited by DLT to 4 characters.
 #[cfg(feature = "dlt-tracing")]
 pub fn new_dlt_tracing<S: tracing_core::Subscriber + for<'a> LookupSpan<'a>>(
-    config: &DltTracingConfig,
+    config: &cda_config::datatypes::DltTracingConfig,
 ) -> Result<BoxedLayer<S>, TracingSetupError> {
     let app_id = tracing_dlt::DltId::try_from(config.app_id.as_str()).map_err(|e| {
         TracingSetupError::ResourceCreationFailed(format!("Invalid DLT app ID: {e}"))
@@ -288,53 +285,4 @@ pub fn init_tracing<T: SubscriberInitExt>(subscriber: T) -> Result<(), TracingSe
             "Failed to initialize tracing subscriber: {e}"
         ))
     })
-}
-
-impl Default for LoggingConfig {
-    fn default() -> Self {
-        Self {
-            timestamp_format: "%T%.3f".to_owned(),
-            log_file_config: LogFileConfig::default(),
-            otel: OtelConfig::default(),
-            #[cfg(feature = "tokio-tracing")]
-            tokio_tracing: TokioTracingConfig::default(),
-            #[cfg(feature = "dlt-tracing")]
-            dlt_tracing: DltTracingConfig::default(),
-        }
-    }
-}
-
-impl Default for LogFileConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            name: DEFAULT_LOG_FILE_NAME.to_owned(),
-            path: DEFAULT_LOG_FILE_PATH.to_owned(),
-            date_format: "%F %T%.3f".to_owned(),
-            append_enabled: false,
-        }
-    }
-}
-
-#[cfg(feature = "tokio-tracing")]
-impl Default for TokioTracingConfig {
-    fn default() -> Self {
-        Self {
-            #[cfg_attr(nightly, allow(unknown_lints, clippy::duration_suboptimal_units))]
-            retention: std::time::Duration::from_secs(60 * 60), // 1h
-            server: "127.0.0.1:6669".to_owned(),
-            recording_path: None,
-        }
-    }
-}
-
-#[cfg(feature = "dlt-tracing")]
-impl Default for DltTracingConfig {
-    fn default() -> Self {
-        Self {
-            app_id: "CDA".to_string(),
-            app_description: "Bridges SOVD to UDS for ECU communication.".to_string(),
-            enabled: true,
-        }
-    }
 }
